# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "spring-saas-core",
    "description": "Control plane multi-tenant da plataforma B2B — tenants, governança ABAC/RBAC, feature flags, auditoria e JWT padronizado.",
    "version": "1.0.0",
    "contact": {
      "name": "Union Solutions"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/audit": {
      "parameters": [
        {
          "$ref": "#/components/parameters/XTenantId"
        },
        {
          "$ref": "#/components/parameters/XCorrelationId"
        }
      ],
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Listar audit logs (cursor pagination)",
        "operationId": "listAuditLogs",
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorSub",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "correlationId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "$ref": "#/components/parameters/cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de audit logs (cursor ou offset dependendo dos params)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CursorPageAudit"
                    },
                    {
                      "$ref": "#/components/schemas/AuditPageResponse"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/v1/audit/export": {
      "parameters": [
        {
          "$ref": "#/components/parameters/XTenantId"
        },
        {
          "$ref": "#/components/parameters/XCorrelationId"
        }
      ],
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Exportar audit logs (compliance)",
        "operationId": "exportAuditLogs",
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ],
              "default": "json"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10000,
              "minimum": 1,
              "maximum": 10000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados exportados (JSON ou CSV conforme formato solicitado)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditExportResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "XTenantId": {
        "name": "X-Tenant-Id",
        "in": "header",
        "description": "ID do tenant (UUID). Opcional quando já presente no JWT.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "XCorrelationId": {
        "name": "X-Correlation-Id",
        "in": "header",
        "description": "ID de correlação para distributed tracing. Gerado automaticamente se ausente.",
        "schema": {
          "type": "string"
        }
      },
      "cursor": {
        "name": "cursor",
        "in": "query",
        "description": "Cursor opaco da página anterior (base64url de Instant)",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Não autenticado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Sem permissão (ABAC)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "AuditLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "actorSub": {
            "type": "string"
          },
          "actorRoles": {
            "type": "string"
          },
          "actorPerms": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "resourceType": {
            "type": "string"
          },
          "resourceId": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "nullable": true
          },
          "correlationId": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "action",
          "createdAt"
        ]
      },
      "CursorPageAudit": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLog"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "hasMore"
        ]
      },
      "AuditPageResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLog"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "content",
          "totalElements"
        ]
      },
      "AuditExportResponse": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date-time"
          },
          "to": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLog"
            }
          }
        },
        "required": [
          "from",
          "to",
          "count",
          "items"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "correlation_id": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      }
    }
  }
}